requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
web3==6.11.3
pyodbc==5.0.1
python-dotenv==1.0.0
ethers==0.1.0
cryptography==41.0.7

